requirements.txt
django==2.1.4
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
python-decouple==3.1
requests==2.22.0
pytz==2018.3
dj-database-url==0.3.0
dj-static==0.0.6
static3==0.6.1
whitenoise==3.3.1
gunicorn==19.7.1
python-dotenv==0.10.3
psycopg2-binary==2.8.3
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
